hardware/raspberrypi/pyproject.toml
[project]
name = "raspberrypi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.57.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rpi-gpio>=0.7.1",
    "smbus>=1.1.post2",
]
